shopify_tracking_automation/requirements.txt
ShopifyAPI
pandas
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
termcolor
